swissairdry/api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
paho-mqtt>=2.1.1
requests>=2.28.2
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.1.0
bcrypt>=4.0.1